added entities for every data or schema
