+++ src/entities/common.ts
@@ -20,3 +20,16 @@ export interface ICoords {
     type: 'Point'
     coordinates: [number, number];
 }
+
+export interface IReview {
+    rating: number
+    review?: string
+    userId: string
+}
+
+export type CancelledBy = 'User' | 'Theater' | 'Admin';
+
+export type RowType = 'A' | 'B' | 'C' | 'D' | 'E' | 'F' | 'G' | 'H' | 'I' | 'J' | 'K' | 'L' | 'M' | 'N' | 'O' | 'P' | 'Q' | 'R' | 'S' | 'T' | 'U';
+export type ColType = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 13 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 | 22 | 23 | 24 | 25 | 26 | 27 | 28 | 29 | 30;
+
+export type PaymentMethod = 'Razorpay' | 'Wallet'
