+++ src/entities/screen.ts
@@ -1,0 +1,9 @@
+export interface IScreen {
+    _id: string
+    theaterId: string
+    name: string
+    rows: string
+    cols: number
+    seatsCount: number
+    seatId: string
+}
