+++ src/entities/show.ts
@@ -1,0 +1,10 @@
+export interface IShow {
+    _id: string
+    movieId: string
+    screenId: string
+    startTime: Date
+    endTime: Date
+    totalSeatCount: number
+    availableSeatCount: number
+    seatId: string
+}
