+++ src/entities/temp/tempTheater.ts
@@ -1,0 +1,19 @@
+import { ICoords, ITheaterAddress } from "../common"
+
+export interface ITempTheater {
+    _id: string
+    name: string
+    email: string
+    otp: number
+    password: string
+    liscenceId: string
+    coords: ICoords,
+    address: ITheaterAddress
+    expireAt: Date
+}
+
+
+// export interface ITempTheaterRes extends ITempTheaterReq {
+//     _id: string
+//     expireAt: Date
+// }
