+++ src/entities/temp/tempUser.ts
@@ -1,0 +1,9 @@
+
+export interface ITempUser {
+    _id: string
+    name: string
+    email: string
+    otp: number
+    password: string
+    expireAt: Date
+}
