+++ src/infrastructure/repositories/movieRepository.ts
@@ -1,7 +1,8 @@
 import { log } from "console";
 import { movieModel } from "../db/movieModel";
 import { IMovieRepo } from "../../useCases/repos/movieRepo";
-import { IMovie, ITMDBMovie } from "../../interfaces/schema/movieSchema";
+import { ITMDBMovie } from "../../interfaces/schema/movieSchema";
+import { IMovie } from "../../entities/movie";
 
 interface IMovieQuery {
     language?: { $in: string[] };
