+++ src/infrastructure/repositories/showRepository.ts
@@ -1,9 +1,10 @@
 import { screenModel } from "../db/screensModel";
 import { showModel } from "../db/showModel";
 import { IShowRepo } from "../../useCases/repos/showRepo";
-import { IMovie } from "../../interfaces/schema/movieSchema";
-import { IScreen } from "../../interfaces/schema/screenSchema";
-import { IShow, IShowRes, IShowToSave, IShowsOnAScreen } from "../../interfaces/schema/showSchema";
+import { IShowRes, IShowToSave, IShowsOnAScreen } from "../../interfaces/schema/showSchema";
+import { IMovie } from "../../entities/movie";
+import { IScreen } from "../../entities/screen";
+import { IShow } from "../../entities/show";
 
 export class ShowRepository implements IShowRepo {
