+++ src/interfaces/schema/adminSchema.ts
@@ -1,12 +1,12 @@
 import { IWalletHistory } from "../../entities/common"
 
-export interface IAdmin {
-    _id: string
-    email: string
-    password: string,
-    wallet: number
-    walletHistory: IWalletHistory[] | []
-}
+// export interface IAdmin {
+//     _id: string
+//     email: string
+//     password: string,
+//     wallet: number
+//     walletHistory: IWalletHistory[] | []
+// }
 
 export interface IAdminRes {
     email: string
