+++ src/interfaces/schema/movieSchema.ts
@@ -1,30 +1,22 @@
+import { IMovie } from "../../entities/movie"
 
-export interface IReview {
-    rating: number
-    review?: string
-    userId: string
-}
 
-export interface IDuration {
-    hours: number,
-    minutes: number
-}
 
-export interface IMovie {
-    _id: string
-    title: string
-    original_title?: string
-    poster_path: string
-    backdrop_path?: string
-    overview: string
-    duration: IDuration
-    language: string
-    tmdbId: number
-    release_date: Date
-    genre_ids: number[]
-    review?: IReview[]
-    isDeleted: boolean
-}
+// export interface IMovie {
+//     _id: string
+//     title: string
+//     original_title?: string
+//     poster_path: string
+//     backdrop_path?: string
+//     overview: string
+//     duration: IDuration
+//     language: string
+//     tmdbId: number
+//     release_date: Date
+//     genre_ids: number[]
+//     review?: IReview[]
+//     isDeleted: boolean
+// }
 
 export interface ITMDBMovie {
     tmdbId: number
