+++ src/interfaces/schema/showSeatsSchema.ts
@@ -1,22 +1,7 @@
-import { RowType } from "../common"
-import { IShowSingleSeat } from "./showSchema"
-
-
-export interface IShowSeatCategory {
-    name: string
-    price: number
-    seats: Map<RowType, IShowSingleSeat[]>
-}
+import { IShowSeatCategory, IShowSeats } from "../../entities/showSeat"
 
 export interface IShowSeatCategoryRes extends IShowSeatCategory {
     // seats: Partial<Record<string, IShowSingleSeat[]>>
-}
-
-export interface IShowSeats {
-    _id: string
-    diamond: IShowSeatCategory
-    gold: IShowSeatCategory
-    silver: IShowSeatCategory
 }
 
 export interface IShowSeatToSave extends Omit<IShowSeats, '_id'> {}
