+++ src/interfaces/schema/theaterSchema.ts
@@ -1,23 +1,23 @@
 import { ICoords, ITheaterAddress, IWalletHistory } from "../../entities/common"
 
-export interface ITheater {
-    _id: string
-    name: string
-    email: string
-    mobile?: number
-    password: string
-    isBlocked: boolean
-    profilePic?: string
-    liscenceId: string
-    screenCount: number
-    wallet: number
-    walletHistory: IWalletHistory[] | []
-    isGoogleAuth: boolean
-    coords: ICoords
-    address: ITheaterAddress,
-    socialMediaHandles?: Map<string, string>
-    approvalStatus: string
-}
+// export interface ITheater {
+//     _id: string
+//     name: string
+//     email: string
+//     mobile?: number
+//     password: string
+//     isBlocked: boolean
+//     profilePic?: string
+//     liscenceId: string
+//     screenCount: number
+//     wallet: number
+//     walletHistory: IWalletHistory[] | []
+//     isGoogleAuth: boolean
+//     coords: ICoords
+//     address: ITheaterAddress,
+//     socialMediaHandles?: Map<string, string>
+//     approvalStatus: string
+// }
 
 export interface ITheaterRes {
     _id: string
